+++ src/main/java/com/pdg/sigma/service/CategoryService.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import com.pdg.sigma.domain.Category;
 
-public interface CategoryService extends GenericService<Category, Long> {
-    List<Category> findByCourseId(Long courseId);
+public interface CategoryService extends GenericService<Category, Integer> {
+    List<Category> findByCourseId(Integer courseId);
 }
